aggregates home components into barrel dir, renders old home page under instutitions tab
