+++ src/js/home/components.jsx
@@ -1,0 +1,4 @@
+export * from './highlights';
+export * from './institutions';
+
+
